[CIR][AArch64] Lower NEON set vector lane intrinsics
